+++ base/sysimg.jl
@@ -27,11 +27,11 @@ let
             :Mmap,
             :Serialization,
             :Libdl,
+            :Printf,
             :Markdown,
             :LibGit2,
             :Logging,
             :Sockets,
-            :Printf,
             :Profile,
             :Dates,
             :DelimitedFiles,
